extensions.py
from flask_sqlalchemy import SQLAlchemy
from flask_jwt_extended import JWTManager

import firebase_admin
from firebase_admin import credentials

import os
import firebase_admin
from firebase_admin import credentials
from dotenv import load_dotenv

load_dotenv()

db = SQLAlchemy()
jwt = JWTManager()

@jwt.token_in_blocklist_loader
def check_if_token_revoked(jwt_header, jwt_payload):
    from models.jwt_blacklist import JWTBlacklist
    return JWTBlacklist.query.filter_by(jti=jwt_payload['jti']).first() is not None

# =========================
# FIREBASE ADMIN INIT
# =========================
if not firebase_admin._apps:
    cred_dict = {
        "type": os.getenv("FIREBASE_TYPE"),
        "project_id": os.getenv("FIREBASE_PROJECT_ID"),
        "private_key_id": os.getenv("FIREBASE_PRIVATE_KEY_ID"),
        "private_key": os.getenv("FIREBASE_PRIVATE_KEY").replace("\\n", "\n"),
        "client_email": os.getenv("FIREBASE_CLIENT_EMAIL"),
        "client_id": os.getenv("FIREBASE_CLIENT_ID"),
        "auth_uri": os.getenv("FIREBASE_AUTH_URI"),
        "token_uri": os.getenv("FIREBASE_TOKEN_URI"),
        "auth_provider_x509_cert_url": os.getenv("FIREBASE_AUTH_PROVIDER_X509_CERT_URL"),
        "client_x509_cert_url": os.getenv("FIREBASE_CLIENT_X509_CERT_URL"),
        "universe_domain": os.getenv("FIREBASE_UNIVERSE_DOMAIN")
    }

    cred = credentials.Certificate(cred_dict)
    firebase_admin.initialize_app(cred)